chore(app): Consolidate state ownership
Locate state access closer to where it is used. Which leaves the chain
crate as a simple provider of data types. One could argue that the state
and its layout is part of the chain definition, so expect further
changes to the layout once the ergonomics become more clear.
