+++ app/src/app_hash.rs
@@ -1,16 +1,10 @@
-use async_trait::async_trait;
-use penumbra_storage::{RootHash, Snapshot};
+use penumbra_storage::RootHash;
 use sha2::{Digest as _, Sha256};
 
 static APPHASH_DOMSEP: &str = "PulzaarAppHash";
 
 #[derive(Clone, Copy, Eq, PartialEq)]
 pub struct AppHash(pub [u8; 32]);
-
-#[async_trait]
-pub trait AppHashRead {
-    async fn app_hash(&self) -> eyre::Result<AppHash>;
-}
 
 impl From<RootHash> for AppHash {
     fn from(value: RootHash) -> Self {
@@ -27,13 +21,5 @@ impl std::fmt::Debug for AppHash {
         f.debug_tuple("AppHash")
             .field(&hex::encode(self.0))
             .finish()
-    }
-}
-
-#[async_trait]
-impl AppHashRead for Snapshot {
-    async fn app_hash(&self) -> eyre::Result<AppHash> {
-        let root = self.root_hash().await.map_err(|err| eyre::eyre!(err))?;
-        Ok(AppHash::from(root))
     }
 }
