+++ app/src/lib.rs
@@ -1,6 +1,10 @@
 mod app;
+mod app_hash;
 mod component;
 mod handler;
+mod state;
 
 pub use app::App;
+pub use app_hash::AppHash;
 pub use component::Component;
+pub use state::{AppHashRead, StateWriteExt};
