+++ chain/Cargo.toml
@@ -13,10 +13,5 @@ publish.workspace       = true
 pulzaar-crypto    = { path = "../crypto" }
 pulzaar-encoding  = { path = "../encoding" }
 
-async-trait       = { workspace = true }
 eyre              = { workspace = true }
-hex               = { workspace = true, features = [ "alloc" ] }
-penumbra-storage  = { workspace = true }
-sha2              = { workspace = true }
 serde             = { workspace = true, features = [ "derive" ] }
-tendermint        = { workspace = true }
