+++ chain/src/lib.rs
@@ -3,21 +3,17 @@
 
 mod account;
 mod amount;
-mod app_hash;
 mod asset;
 mod fee;
 mod funds;
-mod state;
 mod transaction;
 
 pub mod genesis;
 pub mod input;
 
 pub use amount::Amount;
-pub use app_hash::{AppHash, AppHashRead};
 pub use asset::Asset;
 pub use fee::Fee;
 pub use funds::Funds;
 pub use input::Input;
-pub use state::StateWriteExt;
 pub use transaction::Transaction;
