+++ encoding/src/lib.rs
@@ -5,4 +5,4 @@ mod state;
 
 pub use convert::{from_bytes, to_bytes};
 pub use error::Error;
-pub use state::{StateReadBcs, StateWriteBcs};
+pub use state::{StateReadDecode, StateWriteEncode};
